Move XComOperatorLink to separate module
